Move TimeGranularity enum to dbt_semantic_interfaces
